collect/requirements.txt
annotated-types==0.6.0
bcrypt==4.1.3
dnspython==2.6.1
kafka-python==2.0.2
pika==1.3.2
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
sqlmodel==0.0.18
typing_extensions==4.11.0
